re-add rtb.helpers interface to ts definitions
It was removed in c402bfd6afbe9d4708757e8eee16b5e0d27b9a80 but its removal
breaks the build for plugins/plugin-boilerplate.

I have re-added it to make the build pass, but I'm assuming that it was
removed on purpose, so I have marked it as a legacy interface.